fix conversion error in block number
git-svn-id: http://pubsvn.inavare.net/openscada/modules/atlantis/trunk@6336 1b50e8ee-deed-0310-b1de-ab401e1c4d63

+++ org.openscada.da.server.dave/src/org/openscada/da/server/dave/DaveBlockConfigurator.java
@@ -121,7 +121,7 @@ public class DaveBlockConfigurator
 
     private DaveRequestBlock makeBlock ( final BlockConfiguration block )
     {
-        final Request request = new Request ( (byte)block.getArea (), (byte)block.getBlock (), (short)block.getStart (), (short)block.getCount () );
+        final Request request = new Request ( (byte)block.getArea (), (short)block.getBlock (), (short)block.getStart (), (short)block.getCount () );
 
         final DaveRequestBlock deviceBlock = new DaveRequestBlock ( block.getName (), this.device, this.context, request, block.isEnableStatistics (), block.getPeriod () );
         new DaveRequestBlockConfigurator ( this.device.getExecutor (), deviceBlock, block.getType () );
